+++ MetroidDeOuf/Cours04/Bullet.hpp
@@ -24,10 +24,10 @@ public:
 	sf::SoundBuffer*				wallHitSound = nullptr;
 	sf::Sound*						soundOrigin = nullptr;
 
-	Bullet(float _cx, float _cy, int stride, float _size, float _speed, sf::Vector2f _direction = { 0,0 });
+	Bullet(int _cx, int _cy, int stride, float _size, float _speed, sf::Vector2f _direction = { 0,0 });
 	~Bullet();
 
-	void setPosition(float _cx, float _rx, float _cy, float _ry);
+	void setPosition(int _cx, float _rx, int _cy, float _ry);
 	void setDirection(sf::Vector2f dir);
 	void manageMovements(float dt);
 	void syncSprite(float dt);
